+++ packages/EDA/src/eda-tools.ts
@@ -4,10 +4,12 @@ import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
+import {TSNE} from '@keckelt/tsne';
+
 import {_principalComponentAnalysisInWebWorker,
   _partialLeastSquareRegressionInWebWorker} from '../wasm/EDAAPI';
 
-import {checkWasmDimensionReducerInputs, checkUMAPinputs, getRowsOfNumericalColumnns} from './utils';
+import {checkWasmDimensionReducerInputs, checkUMAPinputs, checkTSNEinputs, getRowsOfNumericalColumnns} from './utils';
 
 // Principal components analysis (PCA)
 export async function computePCA(table: DG.DataFrame, features: DG.ColumnList, components: number,
@@ -52,16 +54,8 @@ export async function computeUMAP(features: DG.ColumnList, components: number, e
   // check inputs
   checkUMAPinputs(features, components, epochs, neighbors, minDist, spread);  
 
-  // get ro-by-row data
+  // get row-by-row data
   const data = getRowsOfNumericalColumnns(features);
-
-  /*const options = {
-    nComponents: components,
-    nEpochs: epochs,
-    nNeighbors: neighbors,
-    minDist: minDist,
-    spread: spread
-  };*/
 
   let workerOutput: any;
 
@@ -107,3 +101,58 @@ export async function computeUMAP(features: DG.ColumnList, components: number, e
     DG.Column.fromFloat32Array('UMAP' + i.toString(), umapColumnsData[i])
   ));
 } // computeUMAP
+
+// t-distributed stochastic neighbor embedding (t-SNE)
+export async function computeTSNE(features: DG.ColumnList, components: number, 
+  learningRate: number, perplexity: number, iterations: number): Promise<DG.DataFrame> 
+{
+  // check inputs
+  checkTSNEinputs(features, components, learningRate, perplexity, iterations);
+
+  // get row-by-row data
+  const data = getRowsOfNumericalColumnns(features);
+
+  let workerOutput: any;
+
+  // t-SNE in webworker
+  let promise = new Promise((resolve, reject) => {
+    const worker = new Worker(new URL('workers/tsne-worker.ts', import.meta.url));
+    
+    worker.postMessage({
+      data: data, 
+      options: {
+        learningRate: learningRate,
+        perplexity: perplexity,
+        components: components,
+        iterations: iterations
+    }});
+
+    worker.onmessage = function(e) {
+      worker.terminate();
+      resolve(e.data.embeddings);    
+  }}); 
+  
+  await promise.then(
+    result => { workerOutput = result; },
+    error => { throw new Error ('applying t-SNE fails.'); }
+  );
+
+  const embeddings = workerOutput as any[];
+
+  const rowCount = embeddings.length;
+  const range = [...Array(components).keys()];
+
+  // Create output
+
+  // columns data
+  const umapColumnsData = range.map(_ => new Float32Array(rowCount));  
+
+  // perform transponation
+  for (let i = 0; i < rowCount; ++i)
+    for (let j = 0; j < components; ++j)
+      umapColumnsData[j][i] = embeddings[i][j];
+
+  return DG.DataFrame.fromColumns(range.map(i => 
+    DG.Column.fromFloat32Array('tSNE' + i.toString(), umapColumnsData[i])
+  ));
+} // computeTSNE